+++ NEWS.md
@@ -16,8 +16,9 @@
 - ([#17000](https://github.com/rstudio/rstudio/issues/17000)): Always show .positai folder in Files pane even when set to hide hidden files
 - ([#16069](https://github.com/rstudio/rstudio/issues/16069)): Function auto-completion will now include arguments inherited via @inheritDotParams.
 - ([#16530](https://github.com/rstudio/rstudio/issues/16530)): Added Posit Connect Cloud as a publishing target
-- ([#17330](https://github.com/rstudio/rstudio/issues/17330)): Added Help > Release Notes command to open the RStudio release notes in the browser
 - ([#15360](https://github.com/rstudio/rstudio/issues/15360)): Added `r-max-connections` session option to configure the maximum number of R connections (requires R >= 4.4.0)
+- ([#17330](https://github.com/rstudio/rstudio/issues/17330)): Added Help > Release Notes command to open the RStudio release notes in the browser
+- ([#17344](https://github.com/rstudio/rstudio/issues/17344)): Console process dialogs now support dark theme styling
 
 ### Fixed
 - ([#17005](https://github.com/rstudio/rstudio/issues/17005)): Fixed an issue where the Packages pane was empty when non-PPM repositories were configured alongside PPM repositories and certain packages were installed
@@ -47,6 +48,7 @@
 - ([#17115](https://github.com/rstudio/rstudio/issues/17115)): Fixed an issue on macOS Desktop where the "Close" keyboard shortcut in secondary windows (e.g. Shiny app, Help pop-out) were also processed by the main window
 - ([#14882](https://github.com/rstudio/rstudio/issues/14882)): Fixed an issue where YAML comments in R Markdown documents were dropped or displaced when the front matter was modified (e.g. when toggling "Chunk Output in Console")
 - ([#17333](https://github.com/rstudio/rstudio/issues/17333)): Fixed an issue where the publishing wizard would fail when publishing Quarto documents with Quarto v1.9
+- ([#17353](https://github.com/rstudio/rstudio/issues/17353)): Fixed a crash when the Environment pane inspects S4 objects whose defining package has a broken or missing native DLL
 
 ### Dependencies
 - Ace 1.43.5